README: Updated instructions to use Compose v2 commands
We now require Compose v2. Note that this is the version of the software, not to be confused with the version of the compose schema.

To check your version of compose, run `docker compose version` or `docker-compose version`

To check the version of the compose schema used, check the first line in docker-compose.yml. Currently this is set to 3.9